stream-demux/requirements.txt
confluent-kafka>=2.3.0
pyyaml>=6.0
